// source/src/app/interfaces/VoucherRequest.ts
export interface VoucherRequest {
    codCashier?:                     string;
    codCitySender?:                  string;
    codCountrySender?:               string;
    codCurrencySource?:              string;
    codMTCN?:                        string;
    codOperationTypeNumber?:         string;
    codOperationTypeOperationDesc?:  string;
    codOperatorAgent?:               string;
    codReceiverCity?:                string;
    codTransactionNumber?:           string;
    codTypeIdentificationDeclarant?: string;
    codTypeIdentificationReceiver?:  string;
    codTypeOperationAgent?:          string;
    codTypeTransactionAgent?:        string;
    dateTransaction?:                string;
    foreingExchangeRate?:            string;
    idAgent?:                        string;
    idDeclarant?:                    string;
    idDevice?:                       string;
    idReceiver?:                     string;
    idServicePoint?:                 string;
    informationalText?:              string;
    salesRate?:                      string;
    totalValueTax?:                  string;
    valAddressReceiver?:             string;
    valAutorizedAgent?:              string;
    valAutorizedAgentId?:            string;
    valDeclarantClarification?:      string;
    valDescOperationAgent?:          string;
    valDescTypeOperationAgent?:      string;
    valDescriptionHeaderReceipt?:    string;
    valExpectedCurrency?:            string;
    valExpectedValue?:               string;
    valFirstNameDeclarant?:          string;
    valFirstNameReceiver?:           string;
    valFirstNameSender?:             string;
    valFooter?:                      string;
    valGMF?:                         string;
    valIPDevice?:                    string;
    valLastNameDeclarant?:           string;
    valLastNameReceiver?:            string;
    valLastNameSender?:              string;
    valNameAgent?:                   string;
    valNameServicePoint?:            string;
    valNumeralExchange?:             string;
    valOccupation?:                  string;
    valOperationCurrency?:           string;
    valOperationType?:               string;
    valOperationValue?:              string;
    valPhoneReceiver?:               string;
    valRateExchange?:                string;
    valSourceValue?:                 string;
    valTotalPay?:                    string;
    valTotalTransaction?:            string;
    valTotalTransactionCurrency?:    string;
    valTransactionTime?:             string;
    valUSD?:                         string;
    voucherSize?:                    string;
    reprintNumber?:                  string;
    printerName?:                    string;
    valMiddleNameReceiver?:          string;
    valMiddleLastNameReceiver?:      string;
    valMiddleNameDeclarant?:         string;
    valMiddleLastNameDeclarant?:     string;
    isReprint?:                      string;
    codTypeTx?:                      string;
    fingerprintImage?:               string;
}